Update UDTW23 Document with Final Version Note
This commit marks the UDTW23 document for Old Turkish language (otk) and Old Turkic script (orkh) as its final version. A note has been added to indicate that the document, versioned "202311," will no longer receive updates. Future work on this subject is advised to be handled in a separate document.

Key changes:
- Added a NOTE section to clarify that the document is frozen and will not be updated.
- Specified the version as "202311" for historical and reference purposes.
- Advised the creation of a new document for any further authoring related to the language and script pair.